requirements.txt
appdirs==1.4.3
defusedxml==0.5.0
dj-database-url==0.4.2
Django==1.10.6
django-bootstrap3==8.2.1
djangorestframework==3.6.2
gunicorn==19.7.1
httpie==0.9.9
oauthlib==2.0.2
packaging==16.8
psycopg2==2.7.1
Pygments==2.2.0
PyJWT==1.4.2
pyparsing==2.2.0
python3-openid==3.1.0
requests==2.13.0
requests-oauthlib==0.8.0
six==1.10.0
social-auth-app-django==1.1.0
social-auth-core==1.2.0
whitenoise==3.3.0
